Move actions specific to a authentication and shoppinglist to their own actions files
